fix(main): fix migration

+++ main/migrations/005_add_fields_to_game.sql
@@ -1,3 +1,3 @@
 ALTER TABLE game
-ADD COLUMN noncapture_moves INT DEFAULT 0 NOT NULL,
-ADD COLUMN nonpromoting_moves INT DEFAULT 0 NOT NULL;
+ADD COLUMN noncapture_moves BIGINT DEFAULT 0 NOT NULL,
+ADD COLUMN nonpromoting_moves BIGINT DEFAULT 0 NOT NULL;
